Renamed component to match function name
