+++ src/WebAPI/ClientApp/src/components/ErrorReporter.js
